requirements.txt
numpy>=1.24.0
torch>=2.0.0
eval7>=0.1.8
opencv-python>=4.8.0
mss>=9.0.1
paddleocr>=2.7.0
pytesseract>=0.3.10
pyautogui>=0.9.54
pywinauto>=0.6.8
scikit-learn>=1.3.0
rich>=13.5.0
tensorboard>=2.14.0
pyyaml>=6.0
pillow>=10.0.0
pytest>=7.4.0
